+++ projects/composablekernel/include/ck_tile/ops/gemm/warp/warp_gemm_smfmac_impl.hpp
@@ -3,7 +3,6 @@
 
 #pragma once
 
-#include "ck_tile/core.hpp"
 #include "ck_tile/core/numeric/integer.hpp"
 #include "ck_tile/core/tensor/static_distributed_tensor.hpp"
 namespace ck_tile {
